Split exit in distinct files depending on the grammar
